DRAW-348 refactor: Message Broker 이관
