+++ app/websocket/src/main/kotlin/com/xorker/draw/websocket/broker/WebSocketBroadcaster.kt
@@ -2,11 +2,11 @@ package com.xorker.draw.websocket.broker
 
 import com.xorker.draw.room.RoomId
 import com.xorker.draw.user.UserId
-import com.xorker.draw.websocket.BranchedBroadcastEvent
-import com.xorker.draw.websocket.BroadcastEvent
-import com.xorker.draw.websocket.RespectiveBroadcastEvent
 import com.xorker.draw.websocket.SessionMessage
-import com.xorker.draw.websocket.UnicastEvent
+import com.xorker.draw.websocket.broker.event.BranchedBroadcastEvent
+import com.xorker.draw.websocket.broker.event.BroadcastEvent
+import com.xorker.draw.websocket.broker.event.RespectiveBroadcastEvent
+import com.xorker.draw.websocket.broker.event.UnicastEvent
 import org.springframework.context.ApplicationEventPublisher
 import org.springframework.stereotype.Component
