+++ app/websocket/src/main/kotlin/com/xorker/draw/websocket/broker/event/BroadcastEvent.kt
@@ -1,6 +1,7 @@
-package com.xorker.draw.websocket
+package com.xorker.draw.websocket.broker.event
 
 import com.xorker.draw.room.RoomId
+import com.xorker.draw.websocket.SessionMessage
 
 data class BroadcastEvent(
     val roomId: RoomId,
