+++ app/websocket/src/main/kotlin/com/xorker/draw/websocket/broker/event/RespectiveBroadcastEvent.kt
@@ -1,7 +1,8 @@
-package com.xorker.draw.websocket
+package com.xorker.draw.websocket.broker.event
 
 import com.xorker.draw.room.RoomId
 import com.xorker.draw.user.UserId
+import com.xorker.draw.websocket.SessionMessage
 
 data class RespectiveBroadcastEvent(
     val roomId: RoomId,
