+++ app/websocket/src/main/kotlin/com/xorker/draw/websocket/broker/event/UnicastEvent.kt
@@ -1,6 +1,7 @@
-package com.xorker.draw.websocket
+package com.xorker.draw.websocket.broker.event
 
 import com.xorker.draw.user.UserId
+import com.xorker.draw.websocket.SessionMessage
 
 data class UnicastEvent(
     val userId: UserId,
