--- core/src/main/kotlin/com/xorker/draw/websocket/SessionMessageBroker.kt
@@ -1,8 +1,0 @@
-package com.xorker.draw.websocket
-
-interface SessionMessageBroker {
-    fun unicast(event: UnicastEvent)
-    fun broadcast(event: BroadcastEvent)
-    fun broadcast(event: BranchedBroadcastEvent)
-    fun broadcast(event: RespectiveBroadcastEvent)
-}
